Should fix issue #2
Updated s:ReadTOC such that it allows for the aux-file to be separated from the
main tex file through a build directory.

Added options g:LatexBox_build_dir (and b:build_dir) with short documentation.

Also fixed build dir problem for log files, which allows the log file to be
parsed and the quickfix window to be opened with <leader>le.

Thanks to @hengsku for filereadable -> isdirectory!